Deprecate scenario.img_dir option
The option is deprecated for more than 4 years already and it will
be removed soon, see the change in Tempest [1].

Starting Tempest 25.0.0 release, CONF.scenario.img_file need
a full path for the image. CONF.scenario.img_dir was deprecated
and will be removed in the next release. Till Tempest 25.0.0,
old behavior is maintained and keep working but starting
Tempest 26.0.0, you need to specify the full path in
CONF.scenario.img_file config option.

This patch does analogically the same change as was done in Tempest,
in order to give users of the plugin some time for removing img_dir.

[1] https://review.opendev.org/#/c/710996

Related-Bug: #1393881
Change-Id: I6f1377e5eb9e81c484795fc80341ca87d7b9ed04