Re-enable SSL en p-o-i scenarios 1, 3 and 4
It has been long since we disabled ssl in scenarios. With current
versions of OpenStack projects and libraries, all projects but glance,
which has removed native ssl support with [1], are working fine with
SSL, so we can enable it again.

I'm  leaving scenario002 without SSL to cover both cases in CI.

[1] https://review.opendev.org/c/openstack/glance/+/697969

Change-Id: Id95b73003e1f91bc4bf9cb92b6665c1e6cbe2923