[Ussuri and before] Bump puppet-memcached for tls support
Currently in TripleO we are implementing tls support for memcached,
which requires puppet-memcached >= 3.7.0. This change bumps version of
puppet-memcached used during tets to the one with tls support, so that
we can use the proper version of puppet-memcached in unit tests of
puppet-tripleo.
Because there are no changes made until 3.7.0 which is likely to
break interfaces used in current puppet modules it is considered to
be safe to bump that minor version at this moment.

Related-Bug: #1921716
Change-Id: Ib213e151b92d33d30f4335fcd93f86dbab8f0198