Skip content provider job for irrelevant files
Currently content provider job is not skipped even when tripleo job is
skipped, but this is uselesss.
This change ensures that we rely on the same list of irrelevant files
so that the content provider job is skipped when tripleo job is
skipped.

Change-Id: Iee924f673e3c006606071916c906597cbd891962
(cherry picked from commit cabac611610011a70307f8e96e49fc3bd248b823)