Stable-only: Pin semantic_puppet in CentOS 8
semantic_puppet >= 1.1.0 requires ruby >= 2.7.0 so can no longer be
installed in ruby 2.6.0 in CentOS 8.

Backport note:
This backport removes the version check, because we run no tests in
CentOS 9 in this branch while we run some tests in CentOS 7.
In addition, distro check is also removed, because the same pin is
required in Ubuntu Bionic.

Also, pinning of puppet_forge is removed, because r10k 2.6.4 requires
puppet_forge 2.2.8, which is conflicting with the pin which requires
2.2.6 .

Conflicts:
	install_modules.sh

Closes-Bug: #2016056
Change-Id: Iadff3de808a012c42c2c9a7edb24385c64c7ae78