Add support for for CentOS Stream 9
CentOS has published the official mirror for CentOS Stream 9 [1]. In RDO,
we are working to get ready for it ASAP and we have the required repos
to run p-o-i on CS9 [2].

This patch is adding support to CS9 in p-o-i repo with the required
fixes. Also is adding integrations jobs for scenarios 0-4 and adding
them as non-voting. Note that we don't have promotion pipeline working
so i think it's better to not make it voting yet.

[1] http://mirror.stream.centos.org/9-stream/
[2] https://trunk.rdoproject.org/centos9-master/report.html

Change-Id: I9da46a6aaef3559b24d5a47fb0bde9ae1d09abdb