Train and before: Update TripleO job defenitions
This is manual backport of 12d24fb978070e540d0dcf9e3ea2dac2e6aa079d and
updates tripleo job definitions so that it becomes consistent with ones
in stable/wallaby and later.
Note in stable/train and older we use CentOS7 instead of CentOS8 for
tripleo jobs so host OS version is kept to CentOS7.

Change-Id: I45e1ac9afa2de82d0ad025e7195297240ae45e72