Ubuntu: Use utf8mb3 charset in MySQL
We are observing the failure in Ubuntu integration job caused by
broken idempotency. It seems current mysql/mariadb in Ubuntu replaces
utf8 with uft8mb3 automatically and this behavior results in unexpected
change in second run.

This change ensures that we use utf8mb3 in Ubuntu to avoid such
mismatch caused by internal translation. Note this change is intended
to be a quick gate fix and the charset would be updated to utf8mb4
later.

Closes-Bug: #1929073
Change-Id: I7742120758057b5a00f1d093f355fda3791f4e02