Fix unprivileged ping
Recent iputils in CentOS 8 Stream removed capabilities on the ping
binary and requires the change in net.ipv4.ping_group_range to allow
unprivileged users to ping, however the change in systemd[1] is not
yet shipped.

This is a temporal workaround and sets the kernel parameter. This can
be removed once the fix is shipped as part of a new systemd package.

[1] https://github.com/redhat-plumbers/systemd-rhel8/pull/246

Closes-Bug: #1957913
Change-Id: I6fc0e7a62106c53d983818fc8c8b9d1e1cdebb31