CentOS: Capture yum repository files
When some of the repositories are not accessible, repository list is
not properly captured and that makes it difficult to find out which
repository is wrong or has an issue.
This change ensures the repository configuration files are captured
as part of logs, for easy reference.

Change-Id: I9eae6425761b9a93ab0f44d4cb42d93a1add6b40
(cherry picked from commit 8c220d00d4b72aa6a98852db2453532ec524f179)