Enable Octavia in scenario004
This change enables Octavia in scenario004 integration job so that we
can validate deployment of Octavia by puppet-octavia.
Because of limited resource in CI, noop driver is used and all of
backend operations like amphora instance management are disabled. This
is the same method as is used in a CI job in Octavia itself.

Depends-on: https://review.opendev.org/828063
Change-Id: I627872dd81d5ca576ea33ecf4755bf8de43df76a