Ubuntu: Touch /etc/neutron/fwaas_driver.ini
Currently ubuntu integration jobs are failing. Looking at the error,
it seems l3-agent is not able to start because fwaas_driver.ini is no
longer present. This should be fixed in packaging since neutron-fwaas
was once restored, but until then this change ensures an empty file
is created to avoid that failure.

Change-Id: I6994ad2fd4fd0dbc722e5e0ef1c1609dc68a5044