Skip searching selinux alerts
In integration job, there is a task to search selinux alerts but any
alerts don't effectively make the whole job fail, and we often observe
timeout caused slow sealert command with many denial records.

This change disables the task until we fix all of the current denials.
We capture audit.log in CI so we can review that raw file to find out
any denials.

Change-Id: Ic850acfb2b576224c194ceb361e690c5fbf89838