Nova: Use ssh transport for Ubuntu
Currently integration jobs in Ubuntu are broken because of an issue
with enabling libvirtd socket service.

The issue is caused by the policy in Ubuntu packaging that services
like libvirt or nova-compute are enabled by default.

Until we come up with the solution to workaround the issue, this change
switches migration transport from tcp to ssh.

Change-Id: I3b1839565f1f0ad80a9eb472d1040803ec574815