Replace litmus jobs to CentOS Stream 9 (but non-voting)
Some projects (Oslo, Nova, Cinder, ...) removed support for Python 3.6
and no longer work with CentOS 8 Stream which has Python 3.6.

This change migrates the litmus jobs to CentOS Stream 9 because of that
and teardown of CentOS 8 Stream packages in RDO.

Unfortunately, currently litmus jobs in CentOS Stream 9 are still
broken, because the bolt library does not support Ruby >= 3.0, so
these are kept non-voting. We are still waiting for the fix[1], and
will fix the job as soon as the updated version is released.

[1] https://github.com/puppetlabs/bolt/pull/2987

Change-Id: Ibf4d71f6d6d0cf549175df785b8e7a35e65e618f