Debian/Ubuntu: Fix priority setting to pin ceph packages
Since 13044d9f9e8fe0385ba68a458a301cc82188f08b was merged, we use
the upstream ceph repo to pull packages for Ubuntu, and this broke
the priority setting to avoid installing ceph packages from UCA repo.

This change fixes that and ensure the priority is configured with
the ceph repo.

Change-Id: I7b187d59f169689a7acd5cab013bb74d5cdfbf52