Ubuntu: Migrate jobs from 20.04 (focal) to 22.04 (jammy)
... because the Zed release is available for only 22.04.

Because Ruby 2 is no longer available in Ubuntu 22.04, this drops all
tests with Puppet 6, which supports only Ruby 2. (Ruby 3 is supported
since Puppet 7.8.0)

Also, this disables the separate ceph repository because the upstream
repository does not yet contain packages for Ubuntu Jammy. We install
ceph packages from UCA for the time being but would revisit this later.

Depends-on: https://review.opendev.org/c/openstack/puppet-nova/+/865105
Change-Id: Icbfe677dadb3dab96ea06afe650c524b4abbb753