Ubuntu: Use utf8mb3_general_ci collate in MySQL
Currently idempotency in Ubuntu is broken because of the below change
detected in collate in MySQL.

```
/Stage[main]/Keystone::Db::Mysql/Openstacklib::Db::Mysql[keystone]/
Mysql_database[keystone]/collate: collate changed 'utf8mb3_general_ci'
to 'utf8_general_ci'
```

Similarly to what we observed in the past about charset[1], it seems
MySQL in Ubuntu is automatically converting the collate value and that
is causing the "unexpected" change detected in the 2nd puppet run.

This fixes the idempotency by using utf8mb3_general_ci in Ubuntu to
avoid the mismatch caused by internal translation.

[1] 085d3569021441810c1dff7c8f4396003c690805

Change-Id: I56f31397669d3d7b08aa2e9808947141e003ab0b