litmus: Make sure fixtures directory is cleared up
This is follow up of 3bb7ae978a5db7ae203470a84192f880d23943b5 .

We have seen several failures in the task to upload log files, and this
has been caused by circular symbolic links in spec/fixtures directory.
This change ensures the fixtures directory is cleared after gathering
logs into the logs directory, so that the directory is removed before
the upload task is executed.

Change-Id: I82cdc759ea03c2ad0c2537ec8951dc8a1f1c64ba