Add dns/nameservers parameter to tempest config
After [1], designate-tempest-plugin requires parameter dns.nameservers
in tempest config or it fails some jobs. This patch is setting this
parameter always to <host ip>:5322 as that is what is used when
designate is deployed in p-o-i [2].

I am not doing this parameter conditional as it will just not be used when
designate is not deployed so I guess we do not need to manage that case
separatedly.

[1] https://review.opendev.org/c/openstack/designate-tempest-plugin/+/860116
[2] https://github.com/openstack/puppet-openstack-integration/blob/master/manifests/designate.pp#L92-L97

Change-Id: I0c1fa9c4f7e00aa0c90b996d22dec8c5adef6c48