Make tripleo jobs non-voting
Currently triple jobs are broken in master and zed because of[1].

However the recent announcement[2] implies master and zed of TripleO
will no longer be maintained and we are unsure whether the issue will
be fixed soon.

This makes the TripleO jobs non-voting to unblock gate quickly. We have
to decide whether we should drop these jobs once the decision is made
by the TripleO project and TC.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2169466
[2] https://lists.openstack.org/pipermail/openstack-discuss/2023-February/032083.html

Related-Bug: #2007659
Change-Id: Id300f9a31b106b702ea0ebb890679a5fcc141e10