zed-only: Update bundler to 2.4.22
There is an issue in rubygems that is shipped with Ruby 2.7.0 that
causes bundler install to loop forever when resolving dependencies.
Note that this change is proposed directly to stable/zed because
the pin was removed during 2024.1 cycle.

This also backports 1acedbb7ae26e725422f9080adae6793e7054a37 to fix
lint errors.

Change-Id: Id49d5c16d9e0270b065e655aa92fab4502791cdb