Add non-voting Puppet 8 jobs
Puppet 8 was recently released. This adds a few non-voting jobs to
start testing our modules with that new major version.

Note that Puppet 8 requires Ruby 3.1, which is not available in Ubuntu
Jammy and CentOS 9 Stream by default. To meat that requirement we use
CentOS 9 Stream with the ruby:3.1 module enabled.

Change-Id: I0dc1d3a6c1b40c1e9cce636e9d41ec45e8bbf3ad