Pin puppetlabs-stdlib to 8.6.0
puppetlabs-stdlib 9.0.0 was released recently but it introduced a few
breaking change which the dependent modules have not yet adopted to.
Also a lot of warning messages are observed because of multiple
deprecations.

This pin the module to the previous major version, until we ensure
our modules and the dependent modules have required updates.

Change-Id: I6b8dc28ebc72fef460cb476075d19e0c77be0c48