Neutron: Run neutron api by httpd + mod_wsgi
This updates neutron deployment to use httpd + mod_wsgi to run api.
This is the recommended way to deploy neutron now since the existing
monolithic neutron-server will be removed as part of eventlet removal
initiative.

Note that the change is made only in CentOS because Ubuntu does not
provide the separate service for rpc-server now.

Depends-on: https://review.opendev.org/c/openstack/puppet-ovn/+/933340
Change-Id: I4cd0b9b0ec5a4c98d35a2517d94131e2cdb7e894