Stop capturing /etc/redis.conf
CentOS <= 8 has been using the /etc/redis.conf file, but it was changed
to /etc/redis/redis.conf in CentOS 9.
So we no longer have to check the old path.

This also simplifies the way to capture redis log so that we don't have
to care the actual file name.

Change-Id: I8bf19b002e689e3c7e00f05799927f3433b0c2dc