Keep puppet.log created at repository setup
In integration tests we first run puppet to configure repositories and
then run puppet two times for testing. However we don't move the log
after the first repository configuration, and the file is overridden
by the first deployment run.

This ensures the file created during repository set up is being moved
so that we can review the full records.

Change-Id: Ia3f94d89423bc7c4b9a95f6b4ddb760fdcacfcee