Refactor library jobs
The library syntx job and the library lint job inherit the integration
job template but this pulls a lot of unnecessary repositories. This
change replaces their parent template to reduce the repository
pre-fetched by zuul.

Also, the library unit job base is very similar to the normal unit test
base so can inherit the base one.

Change-Id: If693ef969cea811b490adb908669cf9f08b54682