Undefine ssl certificate parameters for mariadb
After mariadb 10.5.18, mariadb fails to start if ssl cert parameters
are configured in the file but the certificates do not exist [1]. This
problem was fixed in puppetlabs-mysql 13.1.0[2] but we use older
versions (13.0.0 in stable/zed for example) in stable branch jobs.

CentOS Stream 9 has just bumped to mariadb-10.5.22 and this is breaking
all non-ssl jobs.

All releases <= Zed didn't use ssl in mariadb with p-o-i deployments
but puppetlabs-mysql set default values so we need to explicitly
undefine them.

[1] https://mariadb.com/kb/en/mariadb-10-5-18-release-notes/#ssl
[2] https://github.com/puppetlabs/puppetlabs-mysql/issues/1509

Change-Id: Ibe0a00c02dcf855b313ea2e31f098e3d46d2c088