Remove puppet-sahara
The sahara project was marked inactive, because of lack of appropriate
maintenance effort[1] and the project was excluded from Caracal
deliverables. This removes testing with the puppet-sahara module
because we are also retiring this module for the unmaintained project.

[1] https://review.opendev.org/c/openstack/governance/+/899986

Conflicts:
	Puppetfile
	fixtures/scenario003.pp
	manifests/sahara.pp
	zuul.d/base.yaml

Change-Id: Icb639d2724df41035af53e67de7d24558832adb6
(cherry picked from commit 835a008ebfe99b4d0713221d71b6e9548b941493)